.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
pandas
matplotlib
seaborn
reportlab
plotly
python-dotenv
groq
numpy